\name{plotDensity}
\Rdversion{1.1}
\alias{plotDensity}
\title{plotDensity}
\description{
plotDensity
}
\usage{
plotDensity(data, group = rownames(data), xlab = "Abundance")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{ matrix whose ROWS will be plotted as densities}
  \item{group}{ labels for the data }
  \item{xlab}{ plot x-label }
}
%\details{

%}
\value{
}
\references{ ~put references to the literature/web site here ~ }
\author{ Dana Pascovici }
\note{ ~~further notes~~

}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
